requirements.txt
requests>=2.31.0
PyYAML>=6.0
pandas>=2.0.0
textstat>=0.9.0
pytest>=7.0.0
black>=24.0.0
flake8>=6.0.0
